pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stratum-hq"
version = "0.1.0"
description = "Enriched human image dataset pipeline — multi-modal embeddings for any image collection"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "Tim Lawrenz" }]
keywords = ["dataset", "embeddings", "diffusion", "ffhq", "dinov3", "t5", "pose"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "requests",
]

[project.optional-dependencies]
gpu = [
    "transformers>=4.54.0",
    "accelerate>=0.34.0",
    "safetensors>=0.4.0",
]
pose = [
    "onnxruntime>=1.16.0",
    "opencv-python>=4.5.0",
]
publish = [
    "huggingface-hub>=0.20.0",
    "pyarrow>=14.0.0",
]
all = [
    "stratum-hq[gpu,pose,publish]",
]

[project.scripts]
stratum = "stratum.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/stratum"]
